% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeSentiment.R
\name{analyze_sentiment}
\alias{analyze_sentiment}
\title{BBC News article Sentiment analysis}
\usage{
analyze_sentiment(url_end)
}
\arguments{
\item{url_end}{a character string, an ending part of BBC News particular atricle URL (everything after https://www.bbc.com/news/).
For example, article URL is "https://www.bbc.com/news/world-us-canada-51381625".
Only "world-us-canada-51381625" should be pasted}
}
\value{
gg_plot with sentiment distibution across the article (sentiment > 0 positive and blue in color;
        sentiment < 0 negative and red in color)
}
\description{
Analyze sentiment distribution across the article and plot it
}
\details{
Use Corpus with article text without cleaning it as punctuation and stopwords should be preserved.
         Transform Corpus into a dataframe with one row for each article paragraph.
         Calculate sentiment score for each dataframe row and summary statistics for the entire article sentiment (mean, median, min, max).
         Build a dataframe with sentiment disrtibution and plot it
}
\note{
Please, check that URL (url_end) exists before running the function,
      otherwise you will get an "Error in open.connection(x, "rb") : HTTP error 404".
      Please, insert URLs of the articles in English only. Only for BBC News, not BBC Sports , Travel, Worklife, etc.
}
\examples{
analyze_sentiment("world-us-canada-51381625")
analyze_sentiment("entertainment-arts-35232060")
}
