% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeText.R
\name{analyze_text}
\alias{analyze_text}
\title{Analyze BBC News article}
\usage{
analyze_text(url_end, num_word, do_plot = FALSE)
}
\arguments{
\item{url_end}{character string, an ending part of BBC News particular atricle URL (everything after https://www.bbc.com/news/).
For example, article URL is "https://www.bbc.com/news/world-us-canada-51381625".
Only "world-us-canada-51381625" should be pasted}

\item{num_word}{numeric, the number of words to be included in dataframe and plot.
Recommended value for plotting is 50 or 100 (not more than 200) as the plot gets distorted with the larger values
and the word frequency becomes less significant}

\item{do_plot}{logical, if TRUE, wordcloud is rendered. By default plot is not rendered as do_plot = FALSE}
}
\value{
h_df_word_freq - a dataframe with the num_words frequent words in the article (two columns: words and frequency of appearance in the text).
        Optional: a plot of the dataframe is rendered
}
\description{
Build a dataframe with the input number of the most frequent words in the article (word frequency in decreasing order)
             Plot it with a wordcloud (optional)
}
\note{
Please, check that URL (url_end) exists before running the function,
      otherwise you will get an "Error in open.connection(x, "rb") : HTTP error 404".
      Please, insert URLs of the articles in English only. Only for BBC News, not BBC Sports , Travel, Worklife, etc.
}
\examples{
analyze_text("world-us-canada-51381625", 100, FALSE)
analyze_text("entertainment-arts-51398105", 50, TRUE)
}
