% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanText.R
\name{clean_text}
\alias{clean_text}
\title{Clean Corpus with BBC News article}
\usage{
clean_text(url_end)
}
\arguments{
\item{url_end}{character string, an ending part of BBC News particular atricle URL (everything after https://www.bbc.com/news/).
For example, article URL is "https://www.bbc.com/news/world-us-canada-51381625".
Only "world-us-canada-51381625" should be pasted}
}
\value{
art_c - Corpus with cleaned and transformed text documents.
        Corpus represents a collection of text documents with an article text (each article paragraph as a single document in Corpus)
}
\description{
Clean documents in Corpus which include BBC News article text
}
\details{
Transform the symbols like "/", "\\" into space.
         Stem text, remove numbers, remove engligh stopwords, extra spaces and punctuation, convert text to lower case
}
\note{
Please, check that URL (url_end) exists before running the function,
      otherwise you will get an "Error in open.connection(x, "rb") : HTTP error 404".
      Please, insert URLs of articles in English only. Only for BBC News, not BBC Sports , Travel, Worklife, etc.
}
\examples{
clean_text("world-us-canada-51381625")
clean_text("entertainment-arts-35232060")
}
