% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetText.R
\name{get_text}
\alias{get_text}
\title{Get BBC News article}
\usage{
get_text(url_end)
}
\arguments{
\item{url_end}{character string, an ending part of BBC News particular atricle URL (everything after https://www.bbc.com/news/).
For example, article URL is "https://www.bbc.com/news/world-us-canada-51381625".
Only "world-us-canada-51381625" should be pasted}
}
\value{
art_c - Corpus representing a collection of text documents with an article text (each article paragraph as a single document in Corpus)
}
\description{
Scrap article headline and body text from BBC News website (https://www.bbc.com/news/),
             merge them together and create a Corpus for text mining.
}
\note{
Please, check that URL (url_end) exists before running the function,
      otherwise you will get an "Error in open.connection(x, "rb") : HTTP error 404".
      Please, insert URLs of the articles in English only. Only for BBC News, not BBC Sports , Travel, Worklife, etc.
}
\examples{
get_text("world-us-canada-51381625")
get_text("entertainment-arts-35232060")
}
